pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "fac-tools"
version = "0.0.0"
description = "Tools to help manage WP:FA on enwiki"
requires-python = ">= 3.11"
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    "black",
    "bs4",
    "humanize",
    "more-itertools",
    "mwparserfromhell",
    "pytest",
    "pytest-datadir",
    "pywikibot",
    "pytest-socket",
    "requests_oauthlib",
]

[project.scripts]
fac-summary-bot = "fac_tools.scripts.fac_summary_bot:main"
get-fac = "fac_tools.scripts.get_fac:main"
